fix bug with table story knobs
