+++ change_log/next/FE-1719-move_knobs_back_into_table_story.yml
@@ -1,0 +1,1 @@
+Bug Fixes: Moves the `Table` story knobs back into the story file as importing them from a separate file causes them to bleed across other stories in storybook
